+++ app/src/main/java/com/aacademy/homework/data/local/dao/MovieDetailDao.kt
@@ -3,7 +3,7 @@ package com.aacademy.homework.data.local.dao
 import androidx.room.Dao
 import androidx.room.Query
 import androidx.room.Transaction
-import com.aacademy.homework.data.local.model.MovieDetailWithActors
+import com.aacademy.homework.data.model.MovieDetailWithActors
 import io.reactivex.rxjava3.core.Single
 
 @Dao
