+++ app/src/main/java/com/aacademy/homework/data/model/Actor.kt
@@ -1,4 +1,4 @@
-package com.aacademy.homework.data.local.model
+package com.aacademy.homework.data.model
 
 import android.os.Parcel
 import android.os.Parcelable
@@ -7,32 +7,38 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity
-data class Tag(
+data class Actor(
     @PrimaryKey val id: Int,
-    val name: String
+    val firstName: String,
+    val lastName: String,
+    val photoPath: String
 ) : Parcelable {
 
     constructor(parcel: Parcel) : this(
         parcel.readInt(),
+        parcel.readString()!!,
+        parcel.readString()!!,
         parcel.readString()!!
     )
 
     override fun writeToParcel(parcel: Parcel, flags: Int) {
         parcel.writeInt(id)
-        parcel.writeString(name)
+        parcel.writeString(firstName)
+        parcel.writeString(lastName)
+        parcel.writeString(photoPath)
     }
 
     override fun describeContents(): Int {
         return 0
     }
 
-    companion object CREATOR : Creator<Tag> {
+    companion object CREATOR : Creator<Actor> {
 
-        override fun createFromParcel(parcel: Parcel): Tag {
-            return Tag(parcel)
+        override fun createFromParcel(parcel: Parcel): Actor {
+            return Actor(parcel)
         }
 
-        override fun newArray(size: Int): Array<Tag?> {
+        override fun newArray(size: Int): Array<Actor?> {
             return arrayOfNulls(size)
         }
     }
