+++ app/src/main/java/com/aacademy/homework/data/model/MovieDetailWithActors.kt
@@ -1,4 +1,4 @@
-package com.aacademy.homework.data.local.model
+package com.aacademy.homework.data.model
 
 import androidx.room.Embedded
 import androidx.room.Junction
