+++ app/src/main/java/com/aacademy/homework/data/model/Tag.kt
@@ -1,4 +1,4 @@
-package com.aacademy.homework.data.local.model
+package com.aacademy.homework.data.model
 
 import android.os.Parcel
 import android.os.Parcelable
@@ -7,38 +7,32 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity
-data class Actor(
+data class Tag(
     @PrimaryKey val id: Int,
-    val firstName: String,
-    val lastName: String,
-    val photoPath: String
+    val name: String
 ) : Parcelable {
 
     constructor(parcel: Parcel) : this(
         parcel.readInt(),
-        parcel.readString()!!,
-        parcel.readString()!!,
         parcel.readString()!!
     )
 
     override fun writeToParcel(parcel: Parcel, flags: Int) {
         parcel.writeInt(id)
-        parcel.writeString(firstName)
-        parcel.writeString(lastName)
-        parcel.writeString(photoPath)
+        parcel.writeString(name)
     }
 
     override fun describeContents(): Int {
         return 0
     }
 
-    companion object CREATOR : Creator<Actor> {
+    companion object CREATOR : Creator<Tag> {
 
-        override fun createFromParcel(parcel: Parcel): Actor {
-            return Actor(parcel)
+        override fun createFromParcel(parcel: Parcel): Tag {
+            return Tag(parcel)
         }
 
-        override fun newArray(size: Int): Array<Actor?> {
+        override fun newArray(size: Int): Array<Tag?> {
             return arrayOfNulls(size)
         }
     }
