+++ app/src/main/java/com/aacademy/homework/ui/activities/MainActivity.kt
@@ -7,7 +7,7 @@ import androidx.appcompat.widget.Toolbar
 import androidx.core.view.WindowCompat
 import com.aacademy.homework.R.anim
 import com.aacademy.homework.R.id
-import com.aacademy.homework.data.local.model.MoviePreviewWithTags
+import com.aacademy.homework.data.model.MoviePreviewWithTags
 import com.aacademy.homework.databinding.ActivityMainBinding
 import com.aacademy.homework.ui.moviedetail.FragmentMoviesDetails
 import com.aacademy.homework.ui.movielist.FragmentMoviesList
